+++ .github/workflows/container-build.yaml
@@ -61,11 +61,16 @@ jobs:
     # using a lowercase version of the tag since
     # the github environment variables are case-sensitive
     # TODO: add step to run tests inside docker
+    
+    # TODO: currently combining path for nvidia variables 
+    # necessary to launch triton in export container
+    # with path required by train / data projects. 
+    # Should add conditionals in this workflow to set these dynamically
     - 
       name: build docker image
       env:
         tag: ${{ env.REGISTRY }}/${{ github.repository }}/${{ matrix.project }}:${{ github.ref_name }}
-        path: /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
+        path: /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/local/cuda-11.8/bin:/opt/tritonserver/bin:/usr/local/mpi/bin:/usr/local/nvidia/bin:/usr/local/cuda/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/local/ucx/bin
       run: |
         export TAG_LC=${tag,,}
         cat app.tar.gz | docker import --change "ENV PATH=${{ env.path }}" - $TAG_LC
